+++ src/main/java/com/prgrms/rg/domain/ridingpost/application/RidingJoinService.java
@@ -1,11 +1,15 @@
-package com.prgrms.rg.domain.ridingpost.model;
+package com.prgrms.rg.domain.ridingpost.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.prgrms.rg.domain.common.event.DomainEventPublisher;
+import com.prgrms.rg.domain.ridingpost.model.RidingJoinFailException;
+import com.prgrms.rg.domain.ridingpost.model.RidingParticipantRepository;
+import com.prgrms.rg.domain.ridingpost.model.RidingPost;
+import com.prgrms.rg.domain.ridingpost.model.RidingPostFinder;
+import com.prgrms.rg.domain.ridingpost.model.RidingPostNotFoundException;
 import com.prgrms.rg.domain.ridingpost.model.event.RidingJoinEvent;
-import com.prgrms.rg.domain.user.application.NoSuchUserException;
 import com.prgrms.rg.domain.user.application.UserReadService;
 import com.prgrms.rg.domain.user.model.User;
 
@@ -27,7 +31,7 @@ public class RidingJoinService {
 			RidingPost post = postFinder.find(ridingPostId);
 			checkDuplicateJoin(participant, post);
 			post.addParticipant(participant);
-		} catch (NoSuchUserException | RidingPostNotFoundException e) {
+		} catch (RidingPostNotFoundException e) {
 			throw new RidingJoinFailException(e);
 		}
 		RidingJoinEvent event = new RidingJoinEvent(userId, ridingPostId); //addParticipant 내부에서 이벤트 발행하는것을 고려해볼것
